moved problem 69 to trash
